+++ src/main/java/com/atlauncher/workers/InstanceInstaller.java
@@ -2134,7 +2134,10 @@ public class InstanceInstaller extends SwingWorker<Boolean, Void> implements Net
     }
 
     private void installServerBootScripts() throws Exception {
-        if (!isServer) {
+        boolean shouldInstallServerScripts = this.loaderVersion != null
+                && this.loaderVersion.shouldInstallServerScripts();
+
+        if (!isServer || !shouldInstallServerScripts) {
             return;
         }